configured routing, deleted unused files
