--- src/test.js
@@ -1,7 +1,0 @@
-let x = {color: 'blue'}
-y = {name: 'John'}
-console.log(x+y)
-console.log(
-  {
-...x,
-...y})
